CORE: moved all functionality from Composer into Entry
